Ensure the output folder existed for qnn lib-generator
