+++ olive/model/utils/__init__.py
@@ -3,11 +3,12 @@
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
 from olive.model.utils.hf_mappings import HIDDEN_SIZE_NAMES, MODEL_TYPE_MAPPING, NUM_HEADS_NAMES
-from olive.model.utils.onnx_utils import resolve_onnx_path
+from olive.model.utils.model_path_utils import resolve_model_path, resolve_onnx_path
 
 __all__ = [
     "HIDDEN_SIZE_NAMES",
     "MODEL_TYPE_MAPPING",
     "NUM_HEADS_NAMES",
+    "resolve_model_path",
     "resolve_onnx_path",
 ]
