+++ olive/model/utils/onnx_utils.py
@@ -16,30 +16,6 @@ from olive.exception import OliveEvaluationError
 logger = logging.getLogger(__name__)
 
 
-def resolve_onnx_path(file_or_dir_path: str, model_filename: str = "model.onnx") -> str:
-    """Get the model full path.
-
-    The engine provides output paths to ONNX passes that do not contain .onnx extension
-    (these paths are generally locations in the cache). This function will convert such
-    paths to absolute file paths and also ensure the parent directories exist.
-    If the input path is already an ONNX file it is simply returned. Examples:
-
-    resolve_onnx_path("c:/foo/bar.onnx") -> c:/foo/bar.onnx
-
-    resolve_onnx_path("c:/foo/bar") -> c:/foo/bar/model.onnx
-    """
-    if not model_filename.endswith(".onnx"):
-        raise ValueError(f"ONNXModel's model name must end with '.onnx', got {model_filename}")
-
-    path = Path(file_or_dir_path)
-    if path.suffix != ".onnx":
-        path = path / model_filename
-        parent_dir = path.parent
-        if not parent_dir.exists():
-            parent_dir.mkdir(parents=True, exist_ok=True)
-    return str(path)
-
-
 def get_onnx_file_path(model_path: str, onnx_file_name: Optional[str] = None) -> str:
     """Get the path to the ONNX model file.
